backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
PyMySQL==1.1.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0